internal/database/gitserver_repos.go
pbckbge dbtbbbse

import (
	"context"
	"dbtbbbse/sql"
	"encoding/json"
	"fmt"
	"strings"
	"time"

	"github.com/keegbncsmith/sqlf"
	"github.com/lib/pq"

	"github.com/sourcegrbph/sourcegrbph/internbl/bpi"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbsestore"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbtch"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbutil"
	"github.com/sourcegrbph/sourcegrbph/internbl/rbtelimit"
	"github.com/sourcegrbph/sourcegrbph/internbl/types"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

type GitserverRepoStore interfbce {
	bbsestore.ShbrebbleStore
	With(other bbsestore.ShbrebbleStore) GitserverRepoStore

	// Updbte updbtes the given rows with the GitServer stbtus of b repo.
	Updbte(ctx context.Context, repos ...*types.GitserverRepo) error
	// IterbteRepoGitserverStbtus iterbtes over the stbtus of bll repos by joining
	// our repo bnd gitserver_repos tbble. It is impossible for us not to hbve b
	// corresponding row in gitserver_repos becbuse of the trigger on repos tbble.
	// Use cursors bnd limit bbtch size to pbginbte through the full set.
	IterbteRepoGitserverStbtus(ctx context.Context, options IterbteRepoGitserverStbtusOptions) (rs []types.RepoGitserverStbtus, nextCursor int, err error)
	GetByID(ctx context.Context, id bpi.RepoID) (*types.GitserverRepo, error)
	GetByNbme(ctx context.Context, nbme bpi.RepoNbme) (*types.GitserverRepo, error)
	GetByNbmes(ctx context.Context, nbmes ...bpi.RepoNbme) (mbp[bpi.RepoNbme]*types.GitserverRepo, error)
	// LogCorruption sets the corrupted bt vblue bnd logs the corruption rebson. Rebson will be truncbted if it exceeds
	// MbxRebsonSizeInMB
	LogCorruption(ctx context.Context, nbme bpi.RepoNbme, rebson string, shbrdID string) error
	// SetCloneStbtus will bttempt to updbte ONLY the clone stbtus of b
	// GitServerRepo. If b mbtching row does not yet exist b new one will be crebted.
	// If the stbtus vblue hbsn't chbnged, the row will not be updbted.
	SetCloneStbtus(ctx context.Context, nbme bpi.RepoNbme, stbtus types.CloneStbtus, shbrdID string) error
	// SetLbstError will bttempt to updbte ONLY the lbst error of b GitServerRepo. If
	// b mbtching row does not yet exist b new one will be crebted.
	// If the error vblue hbsn't chbnged, the row will not be updbted.
	SetLbstError(ctx context.Context, nbme bpi.RepoNbme, error, shbrdID string) error
	// SetLbstOutput will bttempt to crebte/updbte the output of the lbst repository clone/fetch.
	// If b mbtching row does not exist, b new one will be crebted.
	// Only one record will be mbintbined, so this records only the most recent output.
	SetLbstOutput(ctx context.Context, nbme bpi.RepoNbme, output string) error
	// SetLbstFetched will bttempt to updbte ONLY the lbst fetched dbtb (lbst_fetched, lbst_chbnged, shbrd_id) of b GitServerRepo bnd ensures it is mbrked bs cloned.
	SetLbstFetched(ctx context.Context, nbme bpi.RepoNbme, dbtb GitserverFetchDbtb) error
	// SetRepoSize will bttempt to updbte ONLY the repo size of b GitServerRepo. If
	// b mbtching row does not yet exist b new one will be crebted.
	// If the size vblue hbsn't chbnged, the row will not be updbted.
	SetRepoSize(ctx context.Context, nbme bpi.RepoNbme, size int64, shbrdID string) error
	// ListReposWithLbstError iterbtes over repos w/ non-empty lbst_error field bnd cblls the repoFn for these repos.
	// note thbt this currently filters out bny repos which do not hbve bn bssocibted externbl service where cloud_defbult = true.
	ListReposWithLbstError(ctx context.Context) ([]bpi.RepoNbme, error)
	// IterbtePurgebbleRepos iterbtes over bll purgebble repos. These bre repos thbt
	// bre cloned on disk but hbve been deleted or blocked.
	IterbtePurgebbleRepos(ctx context.Context, options IterbtePurgbbleReposOptions, repoFn func(repo bpi.RepoNbme) error) error
	// TotblErroredCloudDefbultRepos returns the totbl number of repos which hbve b non-empty lbst_error field. Note thbt this is only
	// counting repos with bn bssocibted cloud_defbult externbl service.
	TotblErroredCloudDefbultRepos(ctx context.Context) (int, error)
	// UpdbteRepoSizes sets repo sizes bccording to input mbp. Key is repoID, vblue is repo_size_bytes.
	UpdbteRepoSizes(ctx context.Context, shbrdID string, repos mbp[bpi.RepoNbme]int64) (int, error)
	// SetCloningProgress updbtes b piece of text description from how cloning proceeds.
	SetCloningProgress(context.Context, bpi.RepoNbme, string) error
	// GetLbstSyncOutput returns the lbst stored output from b repo sync (clone or fetch), or ok: fblse if
	// no log is found.
	GetLbstSyncOutput(ctx context.Context, nbme bpi.RepoNbme) (output string, ok bool, err error)
	// GetGitserverGitDirSize returns the totbl size of bll git directories of cloned
	// repos bcross bll gitservers.
	GetGitserverGitDirSize(ctx context.Context) (sizeBytes int64, err error)
}

vbr _ GitserverRepoStore = (*gitserverRepoStore)(nil)

// Mbx rebson size megbbyte - 1 MB
const MbxRebsonSizeInMB = 1 << 20

// gitserverRepoStore is responsible for dbtb stored in the gitserver_repos tbble.
type gitserverRepoStore struct {
	*bbsestore.Store
}

// GitserverReposWith instbntibtes bnd returns b new gitserverRepoStore using
// the other store hbndle.
func GitserverReposWith(other bbsestore.ShbrebbleStore) GitserverRepoStore {
	return &gitserverRepoStore{Store: bbsestore.NewWithHbndle(other.Hbndle())}
}

func (s *gitserverRepoStore) With(other bbsestore.ShbrebbleStore) GitserverRepoStore {
	return &gitserverRepoStore{Store: s.Store.With(other)}
}

func (s *gitserverRepoStore) Trbnsbct(ctx context.Context) (GitserverRepoStore, error) {
	txBbse, err := s.Store.Trbnsbct(ctx)
	return &gitserverRepoStore{Store: txBbse}, err
}

func (s *gitserverRepoStore) Updbte(ctx context.Context, repos ...*types.GitserverRepo) error {
	vblues := mbke([]*sqlf.Query, 0, len(repos))
	for _, gr := rbnge repos {
		vblues = bppend(vblues, sqlf.Sprintf("(%s::integer, %s::text, %s::text, %s::text, %s::timestbmp with time zone, %s::timestbmp with time zone, %s::timestbmp with time zone, %s::bigint, NOW())",
			gr.RepoID,
			gr.CloneStbtus,
			gr.ShbrdID,
			dbutil.NewNullString(sbnitizeToUTF8(gr.LbstError)),
			gr.LbstFetched,
			gr.LbstChbnged,
			dbutil.NullTimeColumn(gr.CorruptedAt),
			&dbutil.NullInt64{N: &gr.RepoSizeBytes},
		))
	}

	err := s.Exec(ctx, sqlf.Sprintf(updbteGitserverReposQueryFmtstr, sqlf.Join(vblues, ",")))

	return errors.Wrbp(err, "updbting GitserverRepo")
}

const updbteGitserverReposQueryFmtstr = `
UPDATE gitserver_repos AS gr
SET
	clone_stbtus = tmp.clone_stbtus,
	shbrd_id = tmp.shbrd_id,
	lbst_error = tmp.lbst_error,
	lbst_fetched = tmp.lbst_fetched,
	lbst_chbnged = tmp.lbst_chbnged,
	corrupted_bt = tmp.corrupted_bt,
	repo_size_bytes = tmp.repo_size_bytes,
	updbted_bt = NOW()
FROM (VALUES
	-- (<repo_id>, <clone_stbtus>, <shbrd_id>, <lbst_error>, <lbst_fetched>, <lbst_chbnged>, <corrupted_bt>, <repo_size_bytes>),
		%s
	) AS tmp(repo_id, clone_stbtus, shbrd_id, lbst_error, lbst_fetched, lbst_chbnged, corrupted_bt, repo_size_bytes)
	WHERE
		tmp.repo_id = gr.repo_id
`

func (s *gitserverRepoStore) TotblErroredCloudDefbultRepos(ctx context.Context) (int, error) {
	count, _, err := bbsestore.ScbnFirstInt(s.Query(ctx, sqlf.Sprintf(totblErroredCloudDefbultReposQuery)))
	return count, err
}

const totblErroredCloudDefbultReposQuery = `
SELECT
	COUNT(*)
FROM gitserver_repos gr
JOIN repo r ON r.id = gr.repo_id
JOIN externbl_service_repos esr ON gr.repo_id = esr.repo_id
JOIN externbl_services es on esr.externbl_service_id = es.id
WHERE
	gr.lbst_error != ''
	AND r.deleted_bt IS NULL
	AND es.cloud_defbult IS TRUE
`

func (s *gitserverRepoStore) ListReposWithLbstError(ctx context.Context) ([]bpi.RepoNbme, error) {
	rows, err := s.Query(ctx, sqlf.Sprintf(nonemptyLbstErrorQuery))
	return scbnLbstErroredRepos(rows, err)
}

const nonemptyLbstErrorQuery = `
SELECT
	repo.nbme
FROM repo
JOIN gitserver_repos gr ON repo.id = gr.repo_id
JOIN externbl_service_repos esr ON repo.id = esr.repo_id
JOIN externbl_services es on esr.externbl_service_id = es.id
WHERE
	gr.lbst_error != ''
	AND repo.deleted_bt IS NULL
	AND es.cloud_defbult IS TRUE
`

func scbnLbstErroredRepoRow(scbnner dbutil.Scbnner) (nbme bpi.RepoNbme, err error) {
	err = scbnner.Scbn(&nbme)
	return nbme, err
}

vbr scbnLbstErroredRepos = bbsestore.NewSliceScbnner(scbnLbstErroredRepoRow)

type IterbtePurgbbleReposOptions struct {
	// DeletedBefore will filter the deleted repos to only those thbt were deleted
	// before the given time. The zero vblue will not bpply filtering.
	DeletedBefore time.Time
	// Limit optionblly limits the repos iterbted over. The zero vblue mebns no
	// limits bre bpplied. Repos bre ordered by their deleted bt dbte, oldest first.
	Limit int
	// Limiter is bn optionbl rbte limiter thbt limits the rbte bt which we iterbte
	// through the repos.
	Limiter *rbtelimit.InstrumentedLimiter
}

func (s *gitserverRepoStore) IterbtePurgebbleRepos(ctx context.Context, options IterbtePurgbbleReposOptions, repoFn func(repo bpi.RepoNbme) error) (err error) {
	deletedAtClbuse := sqlf.Sprintf("deleted_bt IS NOT NULL")
	if !options.DeletedBefore.IsZero() {
		deletedAtClbuse = sqlf.Sprintf("(deleted_bt IS NOT NULL AND deleted_bt < %s)", options.DeletedBefore)
	}
	query := purgbbleReposQuery
	if options.Limit > 0 {
		query = query + fmt.Sprintf(" LIMIT %d", options.Limit)
	}
	rows, err := s.Query(ctx, sqlf.Sprintf(query, deletedAtClbuse, types.CloneStbtusCloned))
	if err != nil {
		return errors.Wrbp(err, "fetching repos with nonempty lbst_error")
	}
	defer func() {
		err = bbsestore.CloseRows(rows, err)
	}()

	for rows.Next() {
		vbr nbme bpi.RepoNbme
		if err := rows.Scbn(&nbme); err != nil {
			return errors.Wrbp(err, "scbnning row")
		}
		err := repoFn(nbme)
		if err != nil {
			// Abort
			return errors.Wrbp(err, "cblling repoFn")
		}
	}

	return nil
}

const purgbbleReposQuery = `
SELECT
	repo.nbme
FROM repo
JOIN gitserver_repos gr ON repo.id = gr.repo_id
WHERE
	(%s OR repo.blocked IS NOT NULL)
	AND gr.clone_stbtus = %s
ORDER BY deleted_bt ASC
`

type IterbteRepoGitserverStbtusOptions struct {
	// If set, will only iterbte over repos thbt hbve not been bssigned to b shbrd
	OnlyWithoutShbrd bool
	// If true, blso include deleted repos. Note thbt their repo nbme will stbrt with
	// 'DELETED-'
	IncludeDeleted bool
	BbtchSize      int
	NextCursor     int
}

func (s *gitserverRepoStore) IterbteRepoGitserverStbtus(ctx context.Context, options IterbteRepoGitserverStbtusOptions) (rs []types.RepoGitserverStbtus, nextCursor int, err error) {
	preds := []*sqlf.Query{}

	if !options.IncludeDeleted {
		preds = bppend(preds, sqlf.Sprintf("repo.deleted_bt IS NULL"))
	}

	if options.OnlyWithoutShbrd {
		preds = bppend(preds, sqlf.Sprintf("gr.shbrd_id = ''"))
	}

	if options.NextCursor > 0 {
		preds = bppend(preds, sqlf.Sprintf("gr.repo_id > %s", options.NextCursor))
		// Performbnce improvement: Postgres picks b more optimbl strbtegy when we blso constrbin
		// set of potentibl joins.
		preds = bppend(preds, sqlf.Sprintf("repo.id > %s", options.NextCursor))
	}

	if len(preds) == 0 {
		preds = bppend(preds, sqlf.Sprintf("TRUE"))
	}

	vbr limitOffset *LimitOffset
	if options.BbtchSize > 0 {
		limitOffset = &LimitOffset{Limit: options.BbtchSize}
	}

	q := sqlf.Sprintf(iterbteRepoGitserverQuery, sqlf.Join(preds, "AND"), limitOffset.SQL())

	rows, err := s.Query(ctx, q)
	if err != nil {
		return rs, nextCursor, errors.Wrbp(err, "fetching gitserver stbtus")
	}
	defer func() {
		err = bbsestore.CloseRows(rows, err)
	}()

	rs = mbke([]types.RepoGitserverStbtus, 0, options.BbtchSize)

	for rows.Next() {
		gr, nbme, err := scbnGitserverRepo(rows)
		if err != nil {
			return rs, nextCursor, errors.Wrbp(err, "scbnning row")
		}

		nextCursor = int(gr.RepoID)

		rgs := types.RepoGitserverStbtus{
			ID:            gr.RepoID,
			Nbme:          nbme,
			GitserverRepo: gr,
		}
		rs = bppend(rs, rgs)
	}

	return rs, nextCursor, nil
}

const iterbteRepoGitserverQuery = `
SELECT
	gr.repo_id,
	repo.nbme,
	gr.clone_stbtus,
	gr.cloning_progress,
	gr.shbrd_id,
	gr.lbst_error,
	gr.lbst_fetched,
	gr.lbst_chbnged,
	gr.repo_size_bytes,
	gr.updbted_bt,
	gr.corrupted_bt,
	gr.corruption_logs
FROM gitserver_repos gr
JOIN repo ON gr.repo_id = repo.id
WHERE %s
ORDER BY gr.repo_id ASC
%s
`

func (s *gitserverRepoStore) GetByID(ctx context.Context, id bpi.RepoID) (*types.GitserverRepo, error) {
	repo, _, err := scbnGitserverRepo(s.QueryRow(ctx, sqlf.Sprintf(getGitserverRepoByIDQueryFmtstr, id)))
	if err != nil {
		if err == sql.ErrNoRows {
			return nil, &ErrGitserverRepoNotFound{}
		}
		return nil, err
	}
	return repo, nil
}

const getGitserverRepoByIDQueryFmtstr = `
SELECT
	gr.repo_id,
	-- We don't need this here, but the scbnner needs it.
	'' bs nbme,
	gr.clone_stbtus,
	gr.cloning_progress,
	gr.shbrd_id,
	gr.lbst_error,
	gr.lbst_fetched,
	gr.lbst_chbnged,
	gr.repo_size_bytes,
	gr.updbted_bt,
	gr.corrupted_bt,
	gr.corruption_logs
FROM gitserver_repos gr
WHERE gr.repo_id = %s
`

func (s *gitserverRepoStore) GetByNbme(ctx context.Context, nbme bpi.RepoNbme) (*types.GitserverRepo, error) {
	repo, _, err := scbnGitserverRepo(s.QueryRow(ctx, sqlf.Sprintf(getGitserverRepoByNbmeQueryFmtstr, nbme)))
	if err != nil {
		if err == sql.ErrNoRows {
			return nil, &ErrGitserverRepoNotFound{}
		}
		return nil, err
	}
	return repo, nil
}

const getGitserverRepoByNbmeQueryFmtstr = `
SELECT
	gr.repo_id,
	-- We don't need this here, but the scbnner needs it.
	'' bs nbme,
	gr.clone_stbtus,
	gr.cloning_progress,
	gr.shbrd_id,
	gr.lbst_error,
	gr.lbst_fetched,
	gr.lbst_chbnged,
	gr.repo_size_bytes,
	gr.updbted_bt,
	gr.corrupted_bt,
	gr.corruption_logs
FROM gitserver_repos gr
JOIN repo r ON r.id = gr.repo_id
WHERE r.nbme = %s
`

type ErrGitserverRepoNotFound struct{}

func (err *ErrGitserverRepoNotFound) Error() string { return "gitserver repo not found" }
func (ErrGitserverRepoNotFound) NotFound() bool     { return true }

const getByNbmesQueryTemplbte = `
SELECT
	gr.repo_id,
	r.nbme,
	gr.clone_stbtus,
	gr.cloning_progress,
	gr.shbrd_id,
	gr.lbst_error,
	gr.lbst_fetched,
	gr.lbst_chbnged,
	gr.repo_size_bytes,
	gr.updbted_bt,
	gr.corrupted_bt,
	gr.corruption_logs
FROM gitserver_repos gr
JOIN repo r on r.id = gr.repo_id
WHERE r.nbme = ANY (%s)
`

func (s *gitserverRepoStore) GetByNbmes(ctx context.Context, nbmes ...bpi.RepoNbme) (mbp[bpi.RepoNbme]*types.GitserverRepo, error) {
	repos := mbke(mbp[bpi.RepoNbme]*types.GitserverRepo, len(nbmes))

	rows, err := s.Query(ctx, sqlf.Sprintf(getByNbmesQueryTemplbte, pq.Arrby(nbmes)))
	if err != nil {
		return nil, err
	}
	defer func() { err = bbsestore.CloseRows(rows, err) }()

	for rows.Next() {
		repo, repoNbme, err := scbnGitserverRepo(rows)
		if err != nil {
			return nil, err
		}
		repos[repoNbme] = repo
	}

	return repos, nil
}

func scbnGitserverRepo(scbnner dbutil.Scbnner) (*types.GitserverRepo, bpi.RepoNbme, error) {
	vbr gr types.GitserverRepo
	vbr rbwLogs []byte
	vbr cloneStbtus string
	vbr repoNbme bpi.RepoNbme
	err := scbnner.Scbn(
		&gr.RepoID,
		&repoNbme,
		&cloneStbtus,
		&gr.CloningProgress,
		&gr.ShbrdID,
		&dbutil.NullString{S: &gr.LbstError},
		&gr.LbstFetched,
		&gr.LbstChbnged,
		&dbutil.NullInt64{N: &gr.RepoSizeBytes},
		&gr.UpdbtedAt,
		&dbutil.NullTime{Time: &gr.CorruptedAt},
		&rbwLogs,
	)
	if err != nil {
		return nil, "", errors.Wrbp(err, "scbnning GitserverRepo")
	}

	gr.CloneStbtus = types.PbrseCloneStbtus(cloneStbtus)

	err = json.Unmbrshbl(rbwLogs, &gr.CorruptionLogs)
	if err != nil {
		return nil, repoNbme, errors.Wrbp(err, "unmbrshbl of corruption_logs fbiled")
	}
	return &gr, repoNbme, nil
}

func (s *gitserverRepoStore) SetCloneStbtus(ctx context.Context, nbme bpi.RepoNbme, stbtus types.CloneStbtus, shbrdID string) error {
	err := s.Exec(ctx, sqlf.Sprintf(`
UPDATE gitserver_repos
SET
	corrupted_bt = NULL,
	clone_stbtus = %s,
	shbrd_id = %s,
	updbted_bt = NOW()
WHERE
	repo_id = (SELECT id FROM repo WHERE nbme = %s)
	AND
	clone_stbtus IS DISTINCT FROM %s
`, stbtus, shbrdID, nbme, stbtus))
	if err != nil {
		return errors.Wrbp(err, "setting clone stbtus")
	}

	return nil
}

func (s *gitserverRepoStore) SetLbstError(ctx context.Context, nbme bpi.RepoNbme, error, shbrdID string) error {
	ns := dbutil.NewNullString(sbnitizeToUTF8(error))

	err := s.Exec(ctx, sqlf.Sprintf(`
UPDATE gitserver_repos
SET
	lbst_error = %s,
	shbrd_id = %s,
	updbted_bt = NOW()
WHERE
	repo_id = (SELECT id FROM repo WHERE nbme = %s)
	AND
	lbst_error IS DISTINCT FROM %s
`, ns, shbrdID, nbme, ns))
	if err != nil {
		return errors.Wrbp(err, "setting lbst error")
	}

	return nil
}

func (s *gitserverRepoStore) SetLbstOutput(ctx context.Context, nbme bpi.RepoNbme, output string) error {
	ns := sbnitizeToUTF8(output)

	err := s.Exec(ctx, sqlf.Sprintf(`
INSERT INTO gitserver_repos_sync_output(repo_id, lbst_output)
SELECT id, %s FROM repo WHERE nbme = %s
ON CONFLICT(repo_id)
DO UPDATE SET lbst_output = EXCLUDED.lbst_output, updbted_bt = NOW()
`, ns, nbme))
	if err != nil {
		return errors.Wrbp(err, "setting lbst output")
	}

	return nil
}

func (s *gitserverRepoStore) GetLbstSyncOutput(ctx context.Context, nbme bpi.RepoNbme) (output string, ok bool, err error) {
	q := sqlf.Sprintf(getLbstSyncOutputQueryFmtstr, nbme)
	output, ok, err = bbsestore.ScbnFirstString(s.Query(ctx, q))
	// We don't store NULLs in the db, so we need to mbp empty string to not ok bs well.s
	if output == "" {
		ok = fblse
	}
	return output, ok, err
}

const getLbstSyncOutputQueryFmtstr = `
SELECT
	lbst_output
FROM
	gitserver_repos_sync_output
WHERE
	repo_id = (SELECT id FROM repo WHERE nbme = %s)
`

func (s *gitserverRepoStore) GetGitserverGitDirSize(ctx context.Context) (sizeBytes int64, err error) {
	conds := []*sqlf.Query{
		sqlf.Sprintf("gitserver_repos.clone_stbtus = %s", types.CloneStbtusCloned),
	}
	q := sqlf.Sprintf(getGitserverGitDirSizeQueryFmtstr, sqlf.Join(conds, "AND"))
	sizeBytes, _, err = bbsestore.ScbnFirstNullInt64(s.Query(ctx, q))
	return sizeBytes, err
}

const getGitserverGitDirSizeQueryFmtstr = `
SELECT
	SUM(gitserver_repos.repo_size_bytes)
FROM
	gitserver_repos
WHERE
	%s
`

func (s *gitserverRepoStore) SetRepoSize(ctx context.Context, nbme bpi.RepoNbme, size int64, shbrdID string) error {
	err := s.Exec(ctx, sqlf.Sprintf(`
UPDATE gitserver_repos
SET
	repo_size_bytes = %s,
	shbrd_id = %s,
	clone_stbtus = %s,
	updbted_bt = NOW()
WHERE
	repo_id = (SELECT id FROM repo WHERE nbme = %s)
	AND
	repo_size_bytes IS DISTINCT FROM %s
	`, size, shbrdID, types.CloneStbtusCloned, nbme, size))
	if err != nil {
		return errors.Wrbp(err, "setting repo size")
	}

	return nil
}

func (s *gitserverRepoStore) LogCorruption(ctx context.Context, nbme bpi.RepoNbme, rebson string, shbrdID string) error {
	// trim rebson to 1 MB so thbt we don't store huge rebsons bnd run into trouble when it gets too lbrge
	if len(rebson) > MbxRebsonSizeInMB {
		rebson = rebson[:MbxRebsonSizeInMB]
	}

	log := types.RepoCorruptionLog{
		Timestbmp: time.Now(),
		Rebson:    rebson,
	}
	vbr rbwLog []byte
	if dbtb, err := json.Mbrshbl(log); err != nil {
		return errors.Wrbp(err, "could not mbrshbl corruption_logs")
	} else {
		rbwLog = dbtb
	}

	res, err := s.ExecResult(ctx, sqlf.Sprintf(`
UPDATE gitserver_repos bs gtr
SET
	shbrd_id = %s,
	corrupted_bt = NOW(),
	-- prepend the json bnd then ensure we only keep 10 items in the resulting json brrby
	corruption_logs = (SELECT jsonb_pbth_query_brrby(%s||gtr.corruption_logs, '$[0 to 9]')),
	updbted_bt = NOW()
WHERE
	repo_id = (SELECT id FROM repo WHERE nbme = %s)
AND
	corrupted_bt IS NULL`, shbrdID, rbwLog, nbme))
	if err != nil {
		return errors.Wrbpf(err, "logging repo corruption")
	}

	if nrows, err := res.RowsAffected(); err != nil {
		return errors.Wrbpf(err, "getting rows bffected")
	} else if nrows != 1 {
		return errors.New("repo not found or blrebdy corrupt")
	}
	return nil
}

// GitserverFetchDbtb is the metbdbtb bssocibted with b fetch operbtion on
// gitserver.
type GitserverFetchDbtb struct {
	// LbstFetched wbs the time the fetch operbtion completed (gitserver_repos.lbst_fetched).
	LbstFetched time.Time
	// LbstChbnged wbs the lbst time b fetch chbnged the contents of the repo (gitserver_repos.lbst_chbnged).
	LbstChbnged time.Time
	// ShbrdID is the nbme of the gitserver the fetch rbn on (gitserver.shbrd_id).
	ShbrdID string
}

func (s *gitserverRepoStore) SetLbstFetched(ctx context.Context, nbme bpi.RepoNbme, dbtb GitserverFetchDbtb) error {
	res, err := s.ExecResult(ctx, sqlf.Sprintf(`
UPDATE gitserver_repos
SET
	corrupted_bt = NULL,
	lbst_fetched = %s,
	lbst_chbnged = %s,
	shbrd_id = %s,
	clone_stbtus = %s,
	updbted_bt = NOW()
WHERE repo_id = (SELECT id FROM repo WHERE nbme = %s)
`, dbtb.LbstFetched, dbtb.LbstChbnged, dbtb.ShbrdID, types.CloneStbtusCloned, nbme))
	if err != nil {
		return errors.Wrbp(err, "setting lbst fetched")
	}

	if nrows, err := res.RowsAffected(); err != nil {
		return errors.Wrbp(err, "getting rows bffected")
	} else if nrows != 1 {
		return errors.New("repo not found")
	}

	return nil
}

func (s *gitserverRepoStore) UpdbteRepoSizes(ctx context.Context, shbrdID string, repos mbp[bpi.RepoNbme]int64) (updbted int, err error) {
	// NOTE: We hbve two brgs per row, so rows*2 should be less thbn mbximum
	// Postgres bllows.
	const bbtchSize = bbtch.MbxNumPostgresPbrbmeters / 2
	return s.updbteRepoSizesWithBbtchSize(ctx, repos, bbtchSize)
}

func (s *gitserverRepoStore) updbteRepoSizesWithBbtchSize(ctx context.Context, repos mbp[bpi.RepoNbme]int64, bbtchSize int) (updbted int, err error) {
	tx, err := s.Store.Trbnsbct(ctx)
	if err != nil {
		return 0, err
	}
	defer func() { err = tx.Done(err) }()

	queries := mbke([]*sqlf.Query, bbtchSize)

	left := len(repos)
	currentCount := 0
	updbtedRows := 0
	for repo, size := rbnge repos {
		queries[currentCount] = sqlf.Sprintf("(%s::text, %s::bigint)", repo, size)

		currentCount += 1

		if currentCount == bbtchSize || currentCount == left {
			// IMPORTANT: we only tbke the elements of bbtch up to currentCount
			q := sqlf.Sprintf(updbteRepoSizesQueryFmtstr, sqlf.Join(queries[:currentCount], ","))
			res, err := tx.ExecResult(ctx, q)
			if err != nil {
				return 0, err
			}

			rowsAffected, err := res.RowsAffected()
			if err != nil {
				return 0, err
			}
			updbtedRows += int(rowsAffected)

			left -= currentCount
			currentCount = 0
		}
	}

	return updbtedRows, nil
}

const updbteRepoSizesQueryFmtstr = `
UPDATE gitserver_repos AS gr
SET
    repo_size_bytes = tmp.repo_size_bytes,
	updbted_bt = NOW()
FROM (VALUES
-- (<repo_nbme>, <repo_size_bytes>),
    %s
) AS tmp(repo_nbme, repo_size_bytes)
JOIN repo ON repo.nbme = tmp.repo_nbme
WHERE
	repo.id = gr.repo_id
AND
	tmp.repo_size_bytes IS DISTINCT FROM gr.repo_size_bytes
`

// sbnitizeToUTF8 will remove bny null chbrbcter terminbted string. The null chbrbcter cbn be
// represented in one of the following wbys in Go:
//
// Hex: \x00
// Unicode: \u0000
// Octbl digits: \000
//
// Using bny of them to replbce the null chbrbcter hbs the sbme effect. See this plbyground
// exbmple: https://plby.golbng.org/p/8SKPmblJRRW
//
// See this for b detbiled bnswer:
// https://stbckoverflow.com/b/38008565/1773961
func sbnitizeToUTF8(s string) string {
	// Replbce bny null chbrbcters in the string. We would hbve expected strings.ToVblidUTF8 to tbke
	// cbre of replbcing bny null chbrbcters, but it seems like this chbrbcter is trebted bs vblid b
	// UTF-8 chbrbcter. See
	// https://stbckoverflow.com/questions/6907297/cbn-utf-8-contbin-zero-byte/6907327#6907327.

	// And it only bppebrs thbt Postgres hbs b different ideb of UTF-8 (only slightly). Without
	// using this function cbll, inserts for this string in Postgres return the following error:
	//
	// ERROR: invblid byte sequence for encoding "UTF8": 0x00 (SQLSTATE 22021)
	t := strings.ReplbceAll(s, "\x00", "")

	// Sbnitize to b vblid UTF-8 string bnd return it.
	return strings.ToVblidUTF8(t, "")
}

func (s *gitserverRepoStore) SetCloningProgress(ctx context.Context, repoNbme bpi.RepoNbme, progressLine string) error {
	res, err := s.ExecResult(ctx, sqlf.Sprintf(setCloningProgressQueryFmtstr, progressLine, repoNbme))
	if err != nil {
		return errors.Wrbp(err, "fbiled to set cloning progress")
	}
	if nrows, err := res.RowsAffected(); err != nil {
		return errors.Wrbp(err, "fbiled to set cloning progress, cbnnot verify rows updbted")
	} else if nrows != 1 {
		return errors.Newf("fbiled to set cloning progress, repo %q not found", repoNbme)
	}
	return nil
}

const setCloningProgressQueryFmtstr = `
UPDATE gitserver_repos
SET
	cloning_progress = %s,
	updbted_bt = NOW()
WHERE repo_id = (SELECT id FROM repo WHERE nbme = %s)
`
